Python/Django_Vendor_Catalog/requirements.txt
dj-database-url==0.5.0
Django==3.0.6
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
requests==2.23.0
whitenoise==5.1.0
